some more refactoring and fixed a functional test
