+++ test/functional/specs/Personalization/C28756.js
@@ -45,6 +45,7 @@ const triggerAlloyEvent = ClientFunction(decisionScope => {
 
 test("Test C28756: A form based offer should return if event command contains its scope.", async () => {
   await configureAlloyInstance("alloy", config);
+  const expectedScopes = ["alloy-test-scope-1", "__view__"];
 
   const result = await triggerAlloyEvent(scope);
 
@@ -59,7 +60,7 @@ test("Test C28756: A form based offer should return if event command contains it
 
   await t
     .expect(requestBody.events[0].query.personalization.decisionScopes)
-    .eql([scope]);
+    .eql(expectedScopes);
 
   const results = [
     "https://ns.adobe.com/personalization/html-content-item",
